Notification icon will show up when users turn off their microphone.
